move options code into separate go file
